+++ CHANGELOG.md
@@ -42,6 +42,8 @@ When releasing, please add the new-release-boilerplate to docs/pallas/CHANGELOG.
 * New features:
   * Added `ResizeMethod.CUBIC_PYTORCH` to {func}`jax.image.resize` to match
     PyTorch's bicubic resize ({jax-issue}`#15768`).
+  * Added {func}`jax.scipy.linalg.qr_multiply` and lower-level 
+    {func}`jax.lax.linalg.qr_multiply` for more efficient QR solves.
   * We now support differentiation of {func}`jax.lax.linalg.qr` for wide
     matrices and when `full_matrices` is `True`.
   * LAPACK operations are now parallelized along the batch dimension on CPU.