+++ CHANGELOG.md
@@ -27,6 +27,8 @@ When releasing, please add the new-release-boilerplate to docs/pallas/CHANGELOG.
     transform matrices ({jax-issue}`#10144`).
   * Added {func}`jax.scipy.linalg.leslie` for constructing Leslie matrices
     ({jax-issue}`#10144`).
+  * Added {func}`jax.scipy.linalg.companion` for constructing companion
+    matrices from polynomial coefficients ({jax-issue}`#10144`).
   * Moved RNG APIs from "implementations" to dtypes ({jax-issue}`#27854`):
     * Added `jax.random.key_dtype` to get the dtype corresponding to a PRNG
       implementation name.