+++ CHANGELOG.md
@@ -31,6 +31,8 @@ When releasing, please add the new-release-boilerplate to docs/pallas/CHANGELOG.
     matrices from polynomial coefficients ({jax-issue}`#10144`).
   * Added {func}`jax.scipy.linalg.fiedler` for constructing symmetric Fiedler
     matrices ({jax-issue}`#10144`).
+  * Added {func}`jax.scipy.linalg.helmert` for constructing Helmert matrices
+    ({jax-issue}`#10144`).
   * Moved RNG APIs from "implementations" to dtypes ({jax-issue}`#27854`):
     * Added `jax.random.key_dtype` to get the dtype corresponding to a PRNG
       implementation name.